+++ src/card/card-section-subtitle/card-section-subtitle.component.ts
@@ -1,0 +1,9 @@
+import { Component, Input, ViewEncapsulation } from '@angular/core';
+
+@Component({
+  selector: 'ign-card-section-subtitle',
+  templateUrl: './card-section-subtitle.html',
+  styleUrls: ['./card-section-subtitle.scss'],
+  encapsulation: ViewEncapsulation.None,
+})
+export class CardSectionSubTitleComponent {}