+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## vNEXT
 * Initial migration of ign components from yamazaki webapp to standalone module
 * Upgrade to Angular 5.x
+* Extract `CardSectionSubTitleComponent`